requirements.txt
asgiref==3.8.1
Django==4.2
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-js-asset==2.2.0
django-taggit==6.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
PyJWT==2.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
